+++ src/__tests__/main.test.ts
@@ -13,40 +13,40 @@ describe("parseAllPacks", () => {
         songs: p.songCount,
       })),
     ).toMatchInlineSnapshot(`
-      Array [
-        Object {
+      [
+        {
           "name": "3rdMix",
           "songs": 38,
         },
-        Object {
+        {
           "name": "5thMix",
           "songs": 40,
         },
-        Object {
+        {
           "name": "A20 (beta)",
           "songs": 101,
         },
-        Object {
+        {
           "name": "Club Fantastic Season 1",
           "songs": 14,
         },
-        Object {
+        {
           "name": "Club Fantastic Season 2",
           "songs": 18,
         },
-        Object {
+        {
           "name": "Dance! @ Anime Destiny 2022",
           "songs": 7,
         },
-        Object {
+        {
           "name": "Easy As Pie 2",
           "songs": 14,
         },
-        Object {
+        {
           "name": "SuperNOVA2",
           "songs": 63,
         },
-        Object {
+        {
           "name": "X2",
           "songs": 44,
         },