+++ src/parseSm.ts
@@ -1,6 +1,6 @@
-import Fraction from "fraction.js";
-import { RawSimfile } from "./parseSong";
-import { FreezeLocation, Arrow } from "./types";
+import { Fraction } from "./fraction.js";
+import { RawSimfile } from "./parseSong.js";
+import { FreezeLocation, Arrow } from "./types.js";
 import {
   determineBeat,
   mergeSimilarBpmRanges,
@@ -8,7 +8,7 @@ import {
   printMaybeError,
   renameBackground,
   reportError,
-} from "./util";
+} from "./util.js";
 
 // Ref: https://github.com/stepmania/stepmania/wiki/sm
 
@@ -193,9 +193,7 @@ export function parseSm(sm: string): RawSimfile {
     const open: Record<number, Partial<FreezeLocation> | undefined> = {};
 
     let curOffset = new Fraction(0);
-    let curMeasureFraction = new Fraction(1).div(
-      getMeasureLength(lines, i) || 1,
-    );
+    let curMeasureFraction = new Fraction(1, getMeasureLength(lines, i) || 1);
 
     for (; i < lines.length && !concludesANoteTag(lines[i]); ++i) {
       const line = lines[i];
@@ -205,7 +203,8 @@ export function parseSm(sm: string): RawSimfile {
       }
 
       if (line[0] === ",") {
-        curMeasureFraction = new Fraction(1).div(
+        curMeasureFraction = new Fraction(
+          1,
           getMeasureLength(lines, i + 1) || 1,
         );
         continue;
@@ -242,7 +241,7 @@ export function parseSm(sm: string): RawSimfile {
               `${sc.title}, ${mode}, ${difficulty} -- error parsing freezes, tried to close a freeze that never opened`,
             );
           } else {
-            const endBeatFraction = curOffset.add(new Fraction(1).div(4));
+            const endBeatFraction = curOffset.add(new Fraction(1, 4));
             thisFreeze.endOffset = endBeatFraction.n / endBeatFraction.d;
             freezes.push(thisFreeze as FreezeLocation);
             open[d] = undefined;
@@ -289,9 +288,7 @@ export function parseSm(sm: string): RawSimfile {
     let curOffset = new Fraction(0);
     // in case the measure is size zero, fall back to dividing by one
     // this is just being defensive, this would mean the stepfile has no notes in it
-    let curMeasureFraction = new Fraction(1).div(
-      getMeasureLength(lines, i) || 1,
-    );
+    let curMeasureFraction = new Fraction(1, getMeasureLength(lines, i) || 1);
 
     for (; i < lines.length && !concludesANoteTag(lines[i]); ++i) {
       // for now, remove freeze ends as they are handled in parseFreezes
@@ -303,7 +300,8 @@ export function parseSm(sm: string): RawSimfile {
       }
 
       if (line.startsWith(",")) {
-        curMeasureFraction = new Fraction(1).div(
+        curMeasureFraction = new Fraction(
+          1,
           getMeasureLength(lines, i + 1) || 1,
         );
         continue;